+++ INASOFT_3.0/Modelos/Devolucion.cs
@@ -12,7 +12,7 @@ namespace INASOFT_3._0.Modelos
 		private string fecha;
 		private string descripcion;
 		private int id_Factura;
-		private int cantidad;
+		private double cantidad;
 		private int id_devolucion;
 		private int id_producto;
 
@@ -20,7 +20,7 @@ namespace INASOFT_3._0.Modelos
         public string Fecha { get => fecha; set => fecha = value; }
         public string Descripcion { get => descripcion; set => descripcion = value; }
         public int Id_Factura { get => id_Factura; set => id_Factura = value; }
-        public int Cantidad { get => cantidad; set => cantidad = value; }
+        public double Cantidad { get => cantidad; set => cantidad = value; }
         public int Id_devolucion { get => id_devolucion; set => id_devolucion = value; }
         public int Id_producto { get => id_producto; set => id_producto = value; }
     }
